pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui-skycoder-tools"
version = "1.0.0"
description = "A collection of useful custom nodes for ComfyUI workflows"
authors = [
    {name = "Skycoder182", email = "skycoder@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["comfyui", "image-processing", "custom-nodes", "ai", "tools", "workflow"]

dependencies = [
    "Pillow>=8.0.0",
    "torch>=1.9.0",
    "numpy>=1.21.0",
]

[project.urls]
Homepage = "https://github.com/skycoder182/comfyui-skycoder-tools"
Repository = "https://github.com/skycoder182/comfyui-skycoder-tools"
Issues = "https://github.com/skycoder182/comfyui-skycoder-tools/issues"
Discussions = "https://github.com/skycoder182/comfyui-skycoder-tools/discussions"

[tool.comfy]
PublisherId = "skycoder"
DisplayName = "Skycoder Tools"
